Update KafkaProducer lib
